add seeders and implement discount for products

+++ database/migrations/2022_09_04_005038_create_products_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->text('introduction')->nullable();
             $table->string('slug');
             $table->string('sku')->nullable();
-            $table->unsignedBigInteger('price');
+            $table->unsignedInteger('price');
             $table->unsignedInteger('rating')->nullable();
             $table->unsignedInteger('quantity')->nullable();
             $table->boolean('is_available')->default(0);
